Separate Pharmanet College API Definitions
