+++ prime-dotnet-webapi/HttpClients/College Licence API/DummyCollegeLicenceClient.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
-using Prime.Models;
+
+using Prime.HttpClients.PharmanetCollegeApiDefinitions;
 
 namespace Prime.HttpClients
 {
