+++ prime-dotnet-webapi/HttpClients/College Licence API/ICollegeLicenceClient.cs
@@ -1,5 +1,6 @@
 using System.Threading.Tasks;
-using Prime.Models;
+
+using Prime.HttpClients.PharmanetCollegeApiDefinitions;
 
 namespace Prime.HttpClients
 {
